add tests for api.VerifyJwt
