+++ api/auth_utils.go
@@ -4,6 +4,7 @@ import (
 	"time"
 
 	"github.com/Ptt-official-app/go-pttbbs/bbs"
+	"gopkg.in/square/go-jose.v2"
 	"gopkg.in/square/go-jose.v2/jwt"
 )
 
@@ -28,4 +29,25 @@ func VerifyJwt(raw string) (userID bbs.UUserID, err error) {
 	}
 
 	return cl.UUserID, nil
+}
+
+func createToken(userec *bbs.Userec) (string, error) {
+	var err error
+
+	sig, err := jose.NewSigner(jose.SigningKey{Algorithm: jose.HS256, Key: JWT_SECRET}, (&jose.SignerOptions{}).WithType("JWT"))
+	if err != nil {
+		return "", err
+	}
+
+	cl := &JwtClaim{
+		UUserID: userec.UUserID,
+		Expire:  jwt.NewNumericDate(time.Now().Add(time.Hour * 72)),
+	}
+
+	raw, err := jwt.Signed(sig).Claims(cl).CompactSerialize()
+	if err != nil {
+		return "", err
+	}
+
+	return raw, nil
 }
