+++ api/login.go
@@ -1,11 +1,7 @@
 package api
 
 import (
-	"time"
-
 	"github.com/Ptt-official-app/go-pttbbs/bbs"
-	"gopkg.in/square/go-jose.v2"
-	"gopkg.in/square/go-jose.v2/jwt"
 )
 
 const LOGIN_R = "/token"
@@ -44,25 +40,4 @@ func Login(remoteAddr string, params interface{}) (interface{}, error) {
 	}
 
 	return result, nil
-}
-
-func createToken(userec *bbs.Userec) (string, error) {
-	var err error
-
-	sig, err := jose.NewSigner(jose.SigningKey{Algorithm: jose.HS256, Key: JWT_SECRET}, (&jose.SignerOptions{}).WithType("JWT"))
-	if err != nil {
-		return "", err
-	}
-
-	cl := &JwtClaim{
-		UUserID: userec.UUserID,
-		Expire:  jwt.NewNumericDate(time.Now().Add(time.Hour * 72)),
-	}
-
-	raw, err := jwt.Signed(sig).Claims(cl).CompactSerialize()
-	if err != nil {
-		return "", err
-	}
-
-	return raw, nil
 }
